requirements.txt
streamlit>=1.32.0
openai>=1.30.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
Pillow>=10.0.0
pandas>=2.0.0
